+++ picodata-plugin/src/error_code.rs
@@ -82,6 +82,9 @@ tarantool::define_enum_with_introspection! {
         /// Picodata machinery is not yet initialized on the instance.
         Uninitialized = 10024,
 
+        /// Attempt to expel an instance which is not currently Offline.
+        ExpelOnlineInstance = 10025,
+
         /// Not an actual error code, just designates the start of the range.
         UserDefinedErrorCodesStart = 20000,
         // Plugin writers should use error codes in this range