+++ test/int/test_replication.py
@@ -263,7 +263,7 @@ def test_expel_blocked_by_replicaset_master_switchover_to_online_replica(
 
     # Initiate master switchover by expelling i4.
     with pytest.raises(CommandFailed) as e:
-        cluster.expel(i4, timeout=1)
+        cluster.expel(i4, timeout=1, force=True)
     assert "Timeout: expel confirmation didn't arrive in time" in e.value.stderr
 
     # Wait until governor switches the replicaset master from i4 to i5
@@ -324,7 +324,7 @@ def test_expel_blocked_by_replicaset_master_switchover_to_offline_replica(
 
     # Initiate master switchover by expelling i4.
     with pytest.raises(CommandFailed) as e:
-        cluster.expel(i4, timeout=1)
+        cluster.expel(i4, timeout=1, force=True)
     assert "Timeout: expel confirmation didn't arrive in time" in e.value.stderr
 
     # Wait until governor switches the replicaset master from i4 to i5